backend/pyproject.toml
[project]
name = "directbnb-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-phonenumber-field[phonenumbers]>=8.3.0",
    "django-storages[s3]>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.4.0",
    "drf-spectacular>=0.27.2",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pyairbnb>=2.1.1",
    "pydantic>=2.12.3",
    "redis>=7.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]
